% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-circ.R
\name{data_circ}
\alias{data_circ}
\title{Add within-period time index column to data.}
\usage{
data_circ(data, period.len = 96, n.obs)
}
\arguments{
\item{data}{CHANGEME}

\item{period.len}{CHANGEME}

\item{n.obs}{CHANGEME}
}
\value{
Matrix: 1st column=within-period time index. 2nd column=data.
}
\description{
Add within-period time index column to data.
}
