% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param-ests.R
\name{param.ests}
\alias{param.ests}
\title{Calculate the estimated parameters in the estimated segments of the (circular time) data}
\usage{
param.ests(
  data,
  period.len = 96,
  params2est = "Mean",
  cpts = c(period.len),
  circdata = TRUE
)
}
\arguments{
\item{data}{CHANGEME}

\item{period.len}{CHANGEME}

\item{params2est}{CHANGEME}

\item{cpts}{CHANGEME}

\item{circdata}{CHANGEME}
}
\value{
estimated parameters
}
\description{
Calculate the estimated parameters in the estimated segments of the (circular time) data
}
