% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_data_gen.R
\name{pgcptSim}
\alias{pgcptSim}
\title{Simulate periodic-global changepoint series (if no global changes, then a time series with stable periodicity is generated)}
\usage{
pgcptSim(
  period = 12,
  dist = "Normal",
  glob.cpts = NA,
  per.cpts = list(),
  params = list(),
  reps = 1
)
}
\arguments{
\item{period}{CHANGEME}

\item{dist}{CHANGEME}

\item{glob.cpts}{CHANGEME}

\item{per.cpts}{CHANGEME}

\item{params}{CHANGEME}

\item{reps}{CHANGEME}
}
\value{
Simulated data
}
\description{
Simulate periodic-global changepoint series (if no global changes, then a time series with stable periodicity is generated)
}
