% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sncirc-bern.R
\name{sncirc.allseg.bern}
\alias{sncirc.allseg.bern}
\title{Calculate all.seg with bernoulli distribution}
\usage{
sncirc.allseg.bern(data, period.len = 96)
}
\description{
Calculate all.seg with bernoulli distribution
}
