% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sncirc-norm.R
\name{sncirc.allseg.norm}
\alias{sncirc.allseg.norm}
\title{Calculate all.seg with normal distribution}
\usage{
sncirc.allseg.norm(data, period.len = 96, dist = "Normal meanvar")
}
\description{
Calculate all.seg with normal distribution
}
