% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sncirc-bern.R
\name{sncirc.bern}
\alias{sncirc.bern}
\title{Run SNcirc with bernoulli distribution}
\usage{
sncirc.bern(
  data,
  period.len = 96,
  max.cpts = 5,
  minseglen = 1,
  pen.val = 0,
  dist = "Bernoulli",
  all.seg
)
}
\description{
Run SNcirc with bernoulli distribution
}
