% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sncirc-norm.R
\name{sncirc.norm}
\alias{sncirc.norm}
\title{Run SNcirc with normal distribution}
\usage{
sncirc.norm(
  data,
  period.len = 96,
  max.cpts = 5,
  minseglen = 1,
  pen.val = 0,
  dist = "Normal meanvar",
  all.seg
)
}
\description{
Run SNcirc with normal distribution
}
