% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sncirc.R
\name{sncirc}
\alias{sncirc}
\title{Run SNcirc on the provided data}
\usage{
sncirc(
  data,
  period.len = 96,
  dist = "Normal meanvar",
  max.cpts = 5,
  minseglen = 1,
  pen.val = 3 * log(length(data)),
  cost = "Likelihood",
  circData = TRUE
)
}
\arguments{
\item{data}{CHANGEME}

\item{period.len}{CHANGEME}

\item{dist}{CHANGEME}

\item{max.cpts}{CHANGEME}

\item{minseglen}{CHANGEME}

\item{pen.val}{CHANGEME}

\item{cost}{CHANGEME}

\item{circData}{CHANGEME}
}
\value{
sncirc results
}
\description{
Run SNcirc on the provided data
}
