% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within-period-plot.R
\name{withinPeriod_plot}
\alias{withinPeriod_plot}
\title{Plot period data}
\usage{
withinPeriod_plot(
  data,
  period.len = 96,
  test.stats = "Normal meanvar",
  title = "",
  param.lines = FALSE,
  cpts = c(period.len),
  params2plot = "Mean",
  circData = TRUE,
  param.est.col = "red",
  defaultAxes = TRUE
)
}
\arguments{
\item{data}{CHANGEME}

\item{test.stats}{CHANGEME}

\item{title}{CHANGEME}

\item{param.lines}{CHANGEME}

\item{cpts}{CHANGEME}

\item{params2plot}{CHANGEME}

\item{circData}{CHANGEME}

\item{param.est.col}{CHANGEME}

\item{defaultAxes}{CHANGEME}

\item{period.ln}{CHANGEME}
}
\description{
Plot period data
}
